requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
Django==3.0.3
djangorestframework==3.11.0
idna==2.9
kiwisolver==1.1.0
matplotlib==3.1.3
numpy==1.18.1
pkg-resources==0.0.0
pylogix==0.5.2
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
